after some discussion with kevin and michael, i have re-implemented the port in a variant-variant scheme rather than main-variant. main only contains key/values that both variants have in common and nothing else.
two variants are provided: stable and unstable. default_variants is set
to +stable. all keys provided in the default variant must be over written
in the others. this behavior is non-ideal and will be revisited pending
on http://opendarwin.org/bugzilla/show_bug.cgi?id=131 additionally one
must provide the master_sites key in main. removal of this is pending on
the resolution of: http://opendarwin.org/bugzilla/show_bug.cgi?id=130

cranked the revision back down to 0 as well


git-svn-id: https://svn.macports.org/repository/macports/trunk@1113 d073be05-634f-4543-b044-5fe20cf6d1d6