removed the grouped contents and now use the super ugly tcl-tastic flavour of contents listing. this does afford using contents-append from the portfile to properly register the two cgi's for removal. disco.
it should also be mentioned that the portfile is semi-gnarly. it utilizes
three reinplaces to deal with the python based interactive install
script. the first reinplace changes the install path, the second
reinplace removes the interactive component of the install and lastly the
third reinplace removes the interactive component of an update install.
interactive python install script?!


git-svn-id: https://svn.macports.org/repository/macports/trunk@1192 d073be05-634f-4543-b044-5fe20cf6d1d6