Bug: Submitted by: Reviewed by: Approved by: Obtained from:
Added the basis for a new FAQ based on bits of the old one written by Jordan.  There are a few issues with this one (aside from the lack of much meaningful content), the css stylesheet needs tweaking and merging with the one used for the Guide, and I've experienced problems with the xsl output.


git-svn-id: https://svn.macports.org/repository/macports/trunk@5204 d073be05-634f-4543-b044-5fe20cf6d1d6