Bug: Submitted by: Reviewed by:    fkr@ Approved by: Obtained from:
port:		mozilla
version:	1.6

This is fkr@'s mozilla port with minor corrections be myself (the
part that breaks is mine).

ok, here it is... aqua variant not wr0king, but this makes mozilla
work with gtk2 (default) and gtk. The reason for this commit (before
the aqua variant is working) is that it makes way for epiphany and
galeon and other stuff that uses mozilla. I have epiphany down and
it should be ready for commiting tomorrow.


git-svn-id: https://svn.macports.org/repository/macports/trunk@5657 d073be05-634f-4543-b044-5fe20cf6d1d6