New port for Cyberduck 2.2 release.
Note: Due to the fact that Apple did not release Java headers with the 1.4.2
Java runtime kit, the Java 1.4.2 SDK (Developer Package) is needed to compile
this. The author has filed a bug with Apple about this, but who knows if they
will correct it. The Portfile detects if it's present and warns and stops the
build if it's not.

Submitted by:   rshaw


git-svn-id: https://svn.macports.org/repository/macports/trunk@5736 d073be05-634f-4543-b044-5fe20cf6d1d6