New port for MacSword 1.0.1
MacSword is a free & open-source application for research and study of God and
His Word. It is developed specifically for Macintosh computers running Mac OS
X.

MacSword allows you to read and browse many different bibles translations in
different languages from Hebrew to Albanian. As well as reading devotionals,
commentaries, dictionaries and lexicons, it also supports searching and
advanced features such as services so that you can access the Bible in any
program.

And because MacSword is based on the Sword project, it is compatible with a
growing collection of over 200 texts available in over 50 languages.

Submitted by:   rshaw


git-svn-id: https://svn.macports.org/repository/macports/trunk@7185 d073be05-634f-4543-b044-5fe20cf6d1d6