Bug: Submitted by: Reviewed by: Approved by: Obtained from:
Since I'm not sure mij@'s original intent with the archives/id.php, I've created an archives.php, which takes id as an argument (using my very rusty and basic PHP knowledge).  So, this is a commit to see if it will work!  I simply made changes to the RSS feed to link to archives.php?id=id.


git-svn-id: https://svn.macports.org/repository/macports/trunk@8672 d073be05-634f-4543-b044-5fe20cf6d1d6