From James' bug report:
- Refactor checksum calculators check_md5 and check_sha1 as calc_md5 and calc_sha1. They now
simply calculate the appropriate type of checksum and return it. The calling code does the actual
checksum verification and thus has the "correct" checksum to report at verbose level.

- Don't bail from checksum checking loop on first error. Thus if there are multiple checksums, all
checksum errors will be reported. This makes it easier/faster to set multiple checksums at once.

----
I did put check on purpose (thinking that there might be a kind of check where it wouldn't be a string comparison), but having all the corrected sums for all the files is cool.

Bug:            #2451
Submitted by:   jberry@
Reviewed by:
Approved by:
Obtained from:


git-svn-id: https://svn.macports.org/repository/macports/trunk@9635 d073be05-634f-4543-b044-5fe20cf6d1d6