Submitted by:   jmpp@ matt@
Various documentation updates, including a new and usefull ports.conf(5) page
and a maybe by now irrelevant portall(1) page. Basically, Matt and I standarized
the dates, .Os clauses and "SEE ALSO" references in all our pages.

I'd like to get some feedback on this commit to see if it can make it into
this round of merges for the next release off of the r1 branch. Particularly
I'd like to hear opinions if at this point is useful having a portall(1) page
(or the command itself for that matter) and making references to it all thoughout,
since users hardly ever walk into the dports/ tree anymore.


git-svn-id: https://svn.macports.org/repository/macports/trunk@13819 d073be05-634f-4543-b044-5fe20cf6d1d6