Revise port(1) to add an expression language that allows groupings of ports to be evaluated.
Using the operators and, or, not, !, (, and ), it is now possible to combine
portnames using using logical expressions.

For instance:

	port echo installed

	...shows all the ports that are installed.

But:

	port echo installed and "apache*"

	...shows all the ports that are installed, and which
	have a name starting with apache.

Likewise:

	port echo "a*" and not \( "apache*" or "auto*" \)

	...escapes ( and ) from the shell, and shows all ports
	starting with "a" that neither start with apache nor with auto.

The words "not" and "!" are equivelant. Note that all tokens must be
passed as individual arguments. The following with not work:

	port echo "a*" and !apache

But this will:

	port echo "a*" and ! apache

Expected operator precedence is respected. Note that straight portnames,
pseudo-portnames, and portname glob patterns may all be intermixed. A straight
sequence or names evaluates the same as if joined by or:

	port echo a b c

...is the same as:

	port echo a or b or c

Interestingly enough, logical groupings of ports can also extended to the
port version/variants. This may or may not prove to be useful.

	port a and a 1.0 ==> a 1.0
	port a 1.0 and a 2.0 ==> <nothing>
	port a 1.0 or a 2.0 ==> a 1.0, a 2.0

Have fun!

James.

p.s. At some point we might further extend this language to allow for
additional port selectors. For instance, the word maintainer:<regex>
might return a list of all ports whose maintainer matches the given
regex.

p.p.s: The registry is slow, and Tcl only a bit more so. These operations
would be sped up by improvements to the registry, such as replacing it with
sqlite. Logical operations on large sets is probably slow due to use of an
interpreted language...etc.


git-svn-id: https://svn.macports.org/repository/macports/trunk@14104 d073be05-634f-4543-b044-5fe20cf6d1d6