Submitted by:   jmpp@
Now that we finally have automated, OD based index regen we no longer need this
file in CVS area, woot!

Thanks to admin@ folks for providing the resources to make this happen!

PS: Please, no one should commit a new index to CVS, as it could clash with the
one generated by the script on dulcinea taking care of the automation.


git-svn-id: https://svn.macports.org/repository/macports/trunk@14352 d073be05-634f-4543-b044-5fe20cf6d1d6