For "port installed foo", make sure we report one way or another whether foo is installed. If it's not installed, state that. Behavior before this change was to give no message for a port that was not installed, giving output only for the installed ports.
This reverts more toward the behavior in 1.1, but I'm not sure it's completely
what we want.

"port installed a b c" might say:

port a is not installed
b 1.2 (active)
c 1.0

"port installed all" would tell you for each port whether it is installed or not.

Feedback welcome. Credit to Juan for reporting this change in behavior.


git-svn-id: https://svn.macports.org/repository/macports/trunk@14667 d073be05-634f-4543-b044-5fe20cf6d1d6