Submitted by:   jmpp@
Here are the files that are meant for the dp dmg, I believe this to be a
better location. However, they also *need* to live in the files/ dir of the
darwinports dport portdir (i.e. dports/sysutils/darwinports/files), so
having them here might be redundant, not too sure yet.


git-svn-id: https://svn.macports.org/repository/macports/trunk@15109 d073be05-634f-4543-b044-5fe20cf6d1d6