Submitted by:   jmpp@ Reviewed by:    jberry@
Reference to the new "Project Membership" chapter of the online guide right
off our front page. This is our response to the very valid points raised on
the "What should I do?" thread on the main user mailing list.

NOTE: The href to the new chapter is explicitly broken for the moment, as I
don't have a clue what the final URL will look like until the XML's are processed
into html files. I'll fix it ASAP once I know.


git-svn-id: https://svn.macports.org/repository/macports/trunk@17430 d073be05-634f-4543-b044-5fe20cf6d1d6