Fix bug #10674.
port installed, active, list, and search all behaved poorly if given
a pseudo port that evaluated to nothing. The root problem was that the
argument parsing code could not tell the difference between no arguments
and arguments that evaluated to a null set.

The change puts a hack in that sets a global option if there are indeed
no arguments to an action, to allow us to distinguish this case.


git-svn-id: https://svn.macports.org/repository/macports/trunk@25789 d073be05-634f-4543-b044-5fe20cf6d1d6