stgit: update to latest version (0.13), with new documentation as variant.
The documentation provided in stgit 0.13 is in asciidoc format, which is
human-readable.  Transforming it into man pages and html format, however,
requires that asciidoc and xmlto be installed, with the latter depending on
libxml2 and libxslt.  Since these extra dependencies are quite large,
building the documentation into those formats is being made into a variant
for now.  Hopefully later releases will include these formats pre-built.


git-svn-id: https://svn.macports.org/repository/macports/trunk@29325 d073be05-634f-4543-b044-5fe20cf6d1d6