doc-new-test: Integrated the man pages more into the guide.
The portfile-* man files are now "normal" xml files for the guide and are
xincluded into it. When the man files are generated the xml are modified so
they are valid man xml files and then generated. With this change the guide
looks better and the structure of the man xml files is clearer.


git-svn-id: https://svn.macports.org/repository/macports/trunk@31178 d073be05-634f-4543-b044-5fe20cf6d1d6