Correct erroneous quoting in environment variables passed to MacPorts. These were written with nested quoting as, e.g, "-L'${prefix}/lib'", but ever since Paul reworked our environment parser back in 1.4.x days and then Anders fixed a bug in the new parser in r30273 (which happened after the 1.5.2 release), the '' subquoting level has been illegal. Now a simple "-L${prefix}/lib", or even -L${prefix}/lib without any quoting at all, suffices.
This commit fixes building and destrooting of bazaar per Adam Mercer (ram)'s bug report on the dev list.
It turned out to not be a regression with respect to 1.5.2 as I originally anticipated, but rather an improper
declaration of the environment, which only worked back in 1.5.2 days as a side effect of the bug Anders fixed
in r30273, if I'm not mistaken.


git-svn-id: https://svn.macports.org/repository/macports/trunk@31808 d073be05-634f-4543-b044-5fe20cf6d1d6